Added gameObject, Physics
I added a gameObject, and a Physics handler. The game object tracks
position, velocity, and image. I added Physics, which takes a refference
to the gameObject Dictionary and then when called will call
gameObject.IntegrateVelocity to move the object, then places the object
in a cell list, last it checks for collisions between objects in the
same cell.